api/v2: fix merge
